+++ waggle-dance-core/src/test/java/com/hotels/bdp/waggledance/mapping/model/DatabaseMappingImplTest.java
@@ -483,7 +483,7 @@ public class DatabaseMappingImplTest {
   @Test
   public void checkWritePermissions() throws Exception {
     databaseMapping.checkWritePermissions(DB_NAME);
-    verify(metastoreMapping).checkWritePermissions(DB_NAME);
+    verify(metastoreMapping).checkWritePermissions(IN_DB_NAME);
   }
 
   @Test